Add files via upload
I made it so the game ends when a player has discarded all of their cards, I also made it cout "UNO!!!" when a player has exactly one card left. Note: the function call to clear the terminal needs to be uncommented out after code is finished, it is currently commented out so it is easier to follow the outputs while debugging.